restructured code moved app file to root and updated imports
